LCD change uint to int

+++ bcl/inc/bc_module_lcd.h
@@ -39,9 +39,9 @@ void bc_module_lcd_init(bc_module_lcd_framebuffer_t *framebuffer);
 void bc_module_lcd_on(void);
 void bc_module_lcd_off(void);
 void bc_module_lcd_clear(void);
-void bc_module_lcd_draw_pixel(uint8_t x, uint8_t y, bool value);
-uint16_t bc_module_lcd_draw_char(uint16_t left, uint16_t top, uint8_t ch);
-uint16_t bc_module_lcd_draw_string(uint16_t left, uint16_t top, char *str);
+void bc_module_lcd_draw_pixel(int x, int y, bool value);
+int bc_module_lcd_draw_char(int left, int top, uint8_t ch);
+int bc_module_lcd_draw_string(int left, int top, char *str);
 //void bc_module_lcd_draw(const uint8_t *frame, uint8_t width, uint8_t height); // In pixels
 //void bc_module_lcd_printf(uint8_t line, /*uint8_t size, font, */const uint8_t *string/*, ...*/);
 void bc_module_lcd_update(void);
